+++ packages/rum/test/utils.ts
@@ -157,30 +157,6 @@ export function parseSegment(bytes: Uint8Array) {
   return JSON.parse(new TextDecoder().decode(bytes)) as BrowserSegment
 }
 
-export function collectAsyncCalls<F extends jasmine.Func>(spy: jasmine.Spy<F>) {
-  return {
-    waitAsyncCalls: (expectedCallsCount: number, callback: (calls: jasmine.Calls<F>) => void) => {
-      if (spy.calls.count() === expectedCallsCount) {
-        callback(spy.calls)
-      } else if (spy.calls.count() > expectedCallsCount) {
-        fail('Unexpected extra call')
-      } else {
-        spy.and.callFake((() => {
-          if (spy.calls.count() === expectedCallsCount) {
-            callback(spy.calls)
-          }
-        }) as F)
-      }
-    },
-    expectNoExtraAsyncCall: (done: () => void) => {
-      spy.and.callFake((() => {
-        fail('Unexpected extra call')
-      }) as F)
-      setTimeout(done, 300)
-    },
-  }
-}
-
 // Returns the first MetaRecord in a Segment, if any.
 export function findMeta(segment: BrowserSegment): MetaRecord | null {
   return segment.records.find((record) => record.type === RecordType.Meta) as MetaRecord
